fix(linux): bundle coherent PipeWire runtime (SCR-441)
